+++ apps/backend/src/modules/user-follows/libs/enums/follow-notification-message.enum.ts.ts
@@ -1,0 +1,5 @@
+const FollowNotificationMessage = {
+	NOTIFICATION: ":firstName :lastName started following you.",
+} as const;
+
+export { FollowNotificationMessage };