+++ apps/backend/src/modules/user-follows/user-follows.service.ts
@@ -1,13 +1,37 @@
+import { configureString } from "~/libs/helpers/helpers.js";
 import { HTTPCode } from "~/libs/modules/http/http.js";
-import { UserFollowsExceptionMessage } from "~/modules/user-follows/libs/enums/enums.js";
+import {
+	FollowNotificationMessage,
+	UserFollowsExceptionMessage,
+} from "~/modules/user-follows/libs/enums/enums.js";
 import { UserFollowsError } from "~/modules/user-follows/libs/exceptions/exceptions.js";
 import { type UserFollowsRepository } from "~/modules/user-follows/user-follows.repository.js";
 
+import {
+	NotificationEntityType,
+	NotificationType,
+} from "../notifications/libs/enums/enums.js";
+import { type NotificationCreateRequestDto } from "../notifications/libs/types/types.js";
+import { type NotificationService } from "../notifications/notification.service.js";
+import { UserExceptionMessage } from "../users/libs/enums/enums.js";
+import { UserError } from "../users/libs/exceptions/exceptions.js";
+import { type UserService } from "../users/users.js";
+
 class UserFollowsService {
+	private notificationService: NotificationService;
+
 	private userFollowsRepository: UserFollowsRepository;
 
-	public constructor(userFollowsRepository: UserFollowsRepository) {
+	private userService: UserService;
+
+	public constructor(
+		notificationService: NotificationService,
+		userFollowsRepository: UserFollowsRepository,
+		userService: UserService,
+	) {
+		this.notificationService = notificationService;
 		this.userFollowsRepository = userFollowsRepository;
+		this.userService = userService;
 	}
 
 	public async follow(
@@ -33,7 +57,34 @@ class UserFollowsService {
 			});
 		}
 
-		return await this.userFollowsRepository.followUser(followerId, followingId);
+		const user = await this.userService.findById(followerId);
+
+		if (!user) {
+			throw new UserError({
+				message: UserExceptionMessage.NOT_FOUND,
+				status: HTTPCode.NOT_FOUND,
+			});
+		}
+
+		const notification: NotificationCreateRequestDto = {
+			content: configureString(FollowNotificationMessage.NOTIFICATION, {
+				firstName: user.firstName,
+				lastName: user.lastName,
+			}),
+			entityId: followerId,
+			entityType: NotificationEntityType.USERS,
+			notificationType: NotificationType.USER_FOLLOWED,
+			userId: followingId,
+		};
+
+		const isFollowed = await this.userFollowsRepository.followUser(
+			followerId,
+			followingId,
+		);
+
+		await this.notificationService.create(notification);
+
+		return isFollowed;
 	}
 
 	public async unfollow(