+++ apps/backend/src/modules/user-routes/libs/types/type.ts
@@ -1,7 +1,7 @@
 export {
 	type UserRouteCreateRequestDto,
-	type UserRouteParameters,
-	type UserRoutePatchRequestDto,
+	type UserRouteFinishRequestDto,
 	type UserRouteResponseDto,
+	type UserRouteStartRequestDto,
 	type UserRouteStatusType,
 } from "@smartscapes/shared";