+++ packages/shared/src/modules/user-routes/libs/enums/user-route-api-path.enum.ts
@@ -1,7 +1,7 @@
 const UserRouteApiPath = {
-	$ID: "/:userId",
-	FINISH: "/:userId/finish",
-	START: "/:userId/start",
+	FINISH: "/finish",
+	ROOT: "/",
+	START: "/start",
 } as const;
 
 export { UserRouteApiPath };