+++ packages/shared/src/modules/user-routes/libs/types/user-route-parameters-request.type.ts
@@ -1,5 +1,0 @@
-type UserRouteParameters = {
-	userId: number;
-};
-
-export { type UserRouteParameters };